Fixed multithreading bug with Name by swapping to Deque + mutex
Deque is better than vector bc as long as we dont erase anything, the memory will never move. The design of the name table is that we NEVER can remove strings from it, therefore a deque is always thread safe for reading.

We still have to lock a mutex for insertion